--- MainLoginScreen.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<!--
-To change this license header, choose License Headers in Project Properties.
-To change this template file, choose Tools | Templates
-and open the template in the editor.
--->
-<html>
-    <head>
-        <title>TODO supply a title</title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    </head>
-    <body>
-        <br/><br/><br/>
-        
-        
-        <form name="login">
-        Username<input type="text" id ="username" name="userid" />
-        Password<input type="password" id ="password" name="pswrd"/><br/>
-        <br/><br/><br/>
-        <input type="button" value="Login" onclick="395login;"/>
-        <input type="reset" value="Reset">
-        </form>
-        
-        
-    </body>
-</html>
