--- inputHours.html
@@ -1,34 +1,0 @@
-<!-- Still a few changes required (moving from xhtml with jsf method to html and java)
-	-basic information fields is simply filler at the moment.
-	-Moving pages after verification -->
-<!DOCTYPE html>
-<html>
-    <header><title>User hours page</title>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	</header>
-
-    <body>
-
-        Basic Information<br/>
-        [Contractor Name: John Doe]<!-- SQL integration  --><br/>
-        [Company Name: Pro Recruiters]<!-- SQL integration  --><br/>
-        [Reporting Period: September 2018]<br/>
-        [Field N:(More information)]<br/>
-	<hr/>
-        Legal Information<br/>
-        By submitting the hours below, you agree that the hours inputted are 
-        correct and hours may be deducted or added depending<br/> on the proper
-        number of hours worked. You also agree to the terms of services as 
-        highlighted in your contract.<!-- Placeholder--></body>
-        <br/></br>
-	<form name="hours">
-            Enter Hours<input type="text" id="hours" name="hours"/>
-            <br/>   
-        <input type="button" value="Submit" onlclick="395submit";">         
-        </form>
-        
-
-    </body>
-</html>
-
