+++ src/pkg395project2019/submitHours.java
@@ -1,6 +1,6 @@
 //Needs SQL integration mainly and method to move to next page.
 
-package project2019;
+package pkg395project2019;
 
 import java.sql.SQLException;
